requirements.txt
certifi==2022.9.24
cfgv==3.3.1
charset-normalizer==2.1.1
distlib==0.3.6
filelock==3.8.0
identify==2.5.5
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.1
nodeenv==1.7.0
pip==22.2.2
platformdirs==2.5.2
pre-commit==2.20.0
python-dotenv==0.21.0
PyYAML==6.0
qbittorrent-api==2022.8.38
requests==2.28.1
setuptools==63.2.0
six==1.16.0
toml==0.10.2
urllib3==1.26.12
virtualenv==20.16.5
